2008-06-23  Michael Snyder  <msnyder@specifix.com>
	* gdbfreeplay.h: Replace all address values with unsigned long long.
	* gdbfreeplay-i386.c: Ditto.
	* gdbfreeplay-mips64.c: Ditto.
	* gdbfreeplay-back.c: Ditto.
	* remote-break.c: Ditto.
	* remote-break.h: Ditto.

+++ gdb/gdbserver/gdbfreeplay.h
@@ -12,12 +12,12 @@ extern int verbose;
 extern enum successcode gdbfreeplay_open (char *filename);
 extern void gdbfreeplay (int socket_fd);
 
-extern unsigned long target_pc_from_T (char *tpacket);
-extern unsigned long target_pc_from_G (char *gpacket);
-extern unsigned long target_pc_from_g (char *gpacket);
+extern unsigned long long target_pc_from_T (char *tpacket);
+extern unsigned long long target_pc_from_G (char *gpacket);
+extern unsigned long long target_pc_from_g (char *gpacket);
 
 extern char *target_compose_T_packet (char *origTpacket, 
-				      unsigned long pc,
+				      unsigned long long pc,
 				      int breakpoint_p);
 extern char *target_compose_g_packet (char *origTpacket);
