+++ gdb/gdbserver/remote-breakpoint.h
@@ -6,7 +6,7 @@
 #define REMOTE_BREAKPOINT_H
 
 typedef struct BREAKPOINT {
-  unsigned long addr;
+  unsigned long long addr;
   unsigned long len;
   struct BREAKPOINT *next;
 } breakpoint;
@@ -31,15 +31,15 @@ enum direction_code {
 };
 
 extern enum successcode remote_remove_breakpoint (enum breakpoint_type,
-						  unsigned long,
+						  unsigned long long,
 						  unsigned long);
 
 extern enum successcode remote_set_breakpoint    (enum breakpoint_type,
-						  unsigned long,
+						  unsigned long long,
 						  unsigned long);
 
 extern enum successcode remote_breakpoint_here_p (enum breakpoint_type,
-						  unsigned long);
+						  unsigned long long);
 
 #endif
